% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_function.R
\name{GET_votation}
\alias{GET_votation}
\title{GET_votation}
\usage{
GET_votation(period = NULL, span = FALSE, party = NULL,
  vote_result = NULL, rows = 500)
}
\arguments{
\item{period}{a scalar or vector of year or years. The period year is defined as the year that starts of fiscal year e.g. 2018 for 2018/19. In vector form it will be e.g. [2018, 2019] will be 2018/19, 2019/20. If @param span is TRUE, then c(2017, 2019) will be evalutated as c(2017, 2018, 2019).}

\item{span}{boolean argument for setting span or not for @param period.}

\item{party}{string with the short names for the parties. E.g. "C" for Centerpartiet.}

\item{vote_result}{string with the possible results from the voting. Possible arguments are Yes, No, Refrain and Absent.}

\item{rows}{integer defining the number of results to return from the query. [0 - 10 000]}
}
\value{
\code{GET_votation} returns a dataframe from the query.
}
\description{
\code{GET_votation} makes a GET request for data from the 
Swedish Parlament API, specifically for votations.
}
\examples{
GET_votation(period=2016, vote_result='Yes', rows=1)
GET_votation(party='C', rows=1)

}
\references{
\url{http://data.riksdagen.se/}
}
