% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rSwedishParliamentVotations}
\alias{rSwedishParliamentVotations}
\alias{rSwedishParliamentVotations-package}
\title{rSwedishParliamentVotations: A linear regression package.}
\description{
The rSwedishParliamentVotations package contains a function \code{GET_votation}, used to query information for votations in the Swedish Parliment.
}
\author{
Laura Julia Melis and Martin Svensson.

Course in Advanced Programming in R.

Department of Computer and Information Science, Linköping University.
}
